requirements.txt
Django
uwsgi
pycountry
psycopg2-binary
django-redis
django_allauth
django-ipware
django_user_agents
geoip2
django-phonenumber-field
phonenumbers
python-dateutil
django-crispy-forms
django-widget-tweaks
bleach
boto3
django-storages==1.8
django-ckeditor
pillow
django-imagekit
python-magic
libmagic
django-recaptcha
python-decouple
requests==2.23.0
numpy
django-ses
